Add LazyEncodingRegistry
A lazy initialization implementation of EncodingRegistry. It does not register any encoding until either the getEncoding(EncodingType) or getEncoding(String) method is called.
When one of these methods is called, the requested EncodingType is registered.
